Refactor sonic command to dedicated module
Move sonic command from osism.commands.manage to osism.commands.sonic
for better code organization. The command maintains all existing
functionality while being isolated in its own module.

AI-assisted: Claude Code
