+++ setup.cfg
@@ -75,7 +75,7 @@ osism.commands:
     manage images = osism.commands.manage:Images
     manage dnsmasq = osism.commands.manage:Dnsmasq
     manage netbox = osism.commands.netbox:Manage
-    manage sonic = osism.commands.manage:Sonic
+    manage sonic = osism.commands.sonic:Manage
     manage redfish list = osism.commands.redfish:List
     manage server list = osism.commands.server:ServerList
     manage server migrate = osism.commands.server:ServerMigrate
